feat: Add Change Password functionality
The code changes include adding a new `ChangePasswordRequest` struct in the `auth_dto.go` file. The `AuthHandler` struct in the `auth_handlers.go` file has been updated to include a `ChangePassword` method that handles the change password functionality. The `AuthRoutes` function in the `auth_routes.go` file has been modified to include a new route for changing the password. The `AuthService` interface and `authService` struct in the `auth_service.go` file have been updated to include a `ChangePassword` method. The `userRepository` struct in the `user_repository.go` file has been modified to include a `ChangePassword` method and an `IsPasswordMatch` method.

This commit adds the functionality to change the password for a user.

```