feat: Add ClassSubject entity and update database connection
The code changes include adding a new `ClassSubject` entity in the `entities` package. The `Connect` function in the `db.go` file has been updated to include the `ClassSubject` entity in the database connection. The `AssignSubjectToClassRequest` and `ClassSubjectResponse` structs have been added to the `subject_dto.go` file.

This commit adds the functionality to assign subjects to classes.