feat: Update teacher subjects response structure
The code changes include updating the `TeacherSubjectsResponse` structure in the `subject_dto.go` file. The `SubjectName` field has been renamed to `Subjects` and its type has been changed to `[]SubjectResponse`. The `TeacherName` field has been moved to the `TeacherSubjectsResponse` structure. This commit improves the structure of the teacher subjects response to provide a more comprehensive and organized representation of the data.