feat: Add StudentID and StudentName fields to AchivementResponse
This commit adds the `StudentID` and `StudentName` fields to the `AchivementResponse` structure in the `achivement_dto.go` file. These fields are necessary to provide information about the student associated with an achievement. The `GetMyAchievements` method in the `StudentHandler` has been updated to populate these fields when retrieving achievements. This commit enhances the functionality of displaying student achievements in the application.