feat: Update UpdateAchivement method in AchivementRepository
The UpdateAchivement method in the AchivementRepository interface has been updated to include an additional parameter, id string, to specify the ID of the achievement to be updated. This change allows for more precise updating of achievements based on their ID. The implementation of the UpdateAchivement method in the achivementRepository struct has also been updated accordingly. This commit improves the functionality of updating achievements in the application.