refactor: Include created and updated timestamps in GetTask response
This commit modifies the `GetTask` function in the `task.go` file of the `StudentHandler` package. It adds the `CreatedAt` and `UpdatedAt` fields to the response struct, allowing the client to access the timestamps of when the task was created and last updated. This change provides more comprehensive information about the tasks and improves the overall functionality of the application.