feat: Add AddQuestion method to QuizRepository and TeacherHandler
This commit adds the AddQuestion method to the QuizRepository interface and implements it in the quizRepository struct. It also adds the AddQuestion handler to the TeacherHandler, allowing teachers to add a new question to a quiz. This feature enhances the functionality of the application by enabling the creation of quizzes with multiple questions.