```text feat: Add SubjectID field to GradeRequest struct
This commit adds the `SubjectID` field to the `GradeRequest` struct in the `grade_dto.go` file. The `SubjectID` field is used to specify the ID of the subject for which the grade is being requested. This modification allows for more accurate and specific grade filtering and retrieval based on the subject.