feat: Add StudentID and TeacherID fields to Parent and Teacher DTOs
This commit adds the `StudentID` field to the `ParentResponse` DTO in the `parent_dto.go` file and the `TeacherID` field to the `TeacherSubjectResponse` DTO in the `subject_dto.go` file. These fields allow for the identification of the student and teacher associated with a parent or subject, respectively. This enhancement improves the completeness and accuracy of the data in the DTOs.