feat: Add FindClassSubjectBySubjectAndClassID and RemoveSubjectFromClass methods to SubjectRepository
This commit adds the FindClassSubjectBySubjectAndClassID and RemoveSubjectFromClass methods to the SubjectRepository interface. The FindClassSubjectBySubjectAndClassID method retrieves a ClassSubject entity based on the given subject ID and class ID. The RemoveSubjectFromClass method removes a subject from a class by deleting the corresponding ClassSubject entity. These methods enhance the functionality of the SubjectRepository and enable better management of class-subject relationships.